requirements.txt
streamlit
python-dotenv
openai
tiktoken
pypdf
faiss-cpu

langchain
langchain-openai
langchain-community
langchain-text-splitters

